+++ src/test/java/com/xiaoleilu/hutool/test/Test.java
@@ -1,11 +1,5 @@
 package com.xiaoleilu.hutool.test;
 
-import java.util.concurrent.TimeUnit;
-
-import com.xiaoleilu.hutool.lang.Console;
-import com.xiaoleilu.hutool.setting.Setting;
-import com.xiaoleilu.hutool.util.ThreadUtil;
-
 /**
  * 仅用于临时测试
  * 
@@ -14,10 +8,6 @@ import com.xiaoleilu.hutool.util.ThreadUtil;
  */
 public class Test {
 	public static void main(String[] args) throws InterruptedException {
-		Setting setting = new Setting("config/db.setting");
-		setting.autoLoad(true);
-		Console.log("####" + setting);
 		
-		ThreadUtil.sleep(5, TimeUnit.HOURS);
 	}
 }